integrated write_egg_data into write_annotate and deleted the base_eeg folder
